+++ t/utils.lisp
@@ -1,10 +1,13 @@
 (defpackage #:weblocks.t.utils
   (:use #:cl
-        #:weblocks)
+        #:weblocks
+        #:hamcrest.prove)
   (:export
    #:with-request
    #:with-session
-   #:is-html))
+   #:is-html
+   #:catch-hooks
+   #:assert-hooks-called))
 (in-package weblocks.t.utils)
 
 
@@ -39,3 +42,35 @@
      (prove:like result ,expected
                ,message)))
 
+
+(defmacro catch-hooks ((&rest hook-names) &body body)
+  "Catches all hook calls and makes available function assert-hooks-contains
+during the body execution. Hook names should be a list of keywords.
+
+Call assert-hooks-contains inside the body, to check if proper hooks were called:
+
+\(handle-hooks \(:fact-created :fact-removed\)
+   \(do-something-to-add-a-fact\)
+   \(assert-hooks-called
+      \(contains :fact-created a-contact a-twitter-name\)
+      \(contains :fact-removed a-contact a-twitter-name\)\)\)"
+  (alexandria:with-gensyms (hook-calls)
+    (let ((hook-handlers
+            (loop for hook-name in hook-names
+                  collect `(weblocks.hooks:add-session-hook
+                               ,hook-name
+                               ,(alexandria:symbolicate 'handle- hook-name)
+                               (&rest args)
+                             (push (cons ,hook-name args)
+                                   ,hook-calls))
+                  )))
+      `(weblocks.hooks:prepare-hooks
+         (let* ((,hook-calls nil)) 
+           ,@hook-handlers
+
+           (macrolet ((assert-hooks-called (&body matchers)
+                        `(assert-that 
+                          ,',hook-calls
+                          (contains
+                           ,@matchers))))
+             ,@body))))))